v1.1 Additions: 1. If method for global condition check 2. If<T> for additional condition checks 3. Rule Priority with new rule builder methods 4. New RuleMetadata class 5. New Fact class to ready for Fact replacing Cell
